+++ common/src/main/java/com/evandev/brick_and_mortar/compat/CompatMods.java
@@ -3,4 +3,5 @@ package com.evandev.brick_and_mortar.compat;
 public class CompatMods {
     public static final String SUPPLEMENTARIES = "supplementaries";
     public static final String ARCHITECTS_PALETTE = "architectspalette";
+    public static final String VANILLA_BACKPORT = "vanillabackport";
 }