+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-06-02
+
+### Added
+
+- Added Vanilla Backport Resin brick variants.
+
+### Fixed
+
+- Fixed missing item tags for stairs, slabs, and walls.
+- Hopefully fixed crash when using the kiln menu with Sinytra.
+
 ## [1.0.1] - 2026-05-30
 
 ### Changed